beam_ssa: Add utility for duplicating Phi values
When adding a predecessor to a node in which there are Phi
instructions, where the added predecessor is itself a successor of one
of the existing predecessors, the Phi instructions need to be
extended. This patch introduces an utility function for doing this
transformation. Consider the following example:

```
a:
  ...
  br b

b:
  x = phi(x0: a, ...)
```

If we insert a `c` as an alternative extended path from `a` to `b`:

```
a:
  ...
  br <cond>, b, c

b:
  x = phi(x0: a...)

c:
  ...
  br b
```

We have to provide a value for x, for the case when we come from `c`,
a call to copy_phi_values(`a`, `b`, `c`, ...) will give us:

```
a:
  ...
  br <cond>, b, c

b:
  x = phi(x0: a, x0:c, ...)

c:
  ...
  br b
```