beam_ssa_opt: Convert bool ops to flow-control
Rewrite boolean bif:'and', bif:'and' and bif:'not' to use flow
control, in effect this optimization convert 'and' and 'or' to be
short circuiting when it can be done safely.

We look for basic blocks which end with a conditional branch branching
on the result of the boolean bif. We only consider bifs which are not
followed by a succeeded instruction and where the conditional branch
is the only user.

Moving of side-effect-free operations to after the first conditional
branch is left for a later pass.