Add ssa_opt_bool_bifs_to_fc2 pass
The ssa_opt_bool_bifs_to_fc2 pass has a wider scope than
ssa_opt_bool_bifs_to_fc in that it looks at all boolean and/or
expressions and tries to convert them to flow-control. What it does is
essentially:

1:
  C=bif:'and' A, B

is rewritten to:

1:
  br A, 2, 3

2:
  br 3

3:
  C = phi 1: false, 2: B

a similar transform is done for 'or' operations.

TODO: Consider if this pass should only trigger when
ssa_opt_bool_bifs_to_fc fails to do any transformations.