beam_ssa_opt: Short circuit failing path of boolean bifs in guards
Try to short circuit the failing path of bif:'and', bif:'or', and
bif:'not' in guards.

We detect the pattern:

Predecessor:
  br/switch to Next

Next:
  X = phi {{literal, Lit}, Predecessor}, ...
  ...
  A = bif:<unsafe bif>(..., X, ...)
  B = succeeded,guard A
  br B, Success, Fail

If Lit is not 'true' nor 'false' we know that the succeed check
will fail, and thus change the branch from Predecessor to Next to
branch directly to Fail. This short circuit is safe if all the
following conditions apply:

* The live out set of Predecessor is a superset of the live-in set
  of the Fail block.

* All execution paths from Predecessor reach the unsafe bif or the
  Fail block.

* All execution paths from Predecessor reaching the unsafe bif or
  branching to the Fail block are free of side-effects.