beam_ssa_opt: Short circuit failing path of boolean bifs in guards
Try to short circuit the failing path of bif:'and', bif:'or',
bif:'not' in guards. We detect the pattern:

P:
  br/switch to N

N:
  X = phi {{literal, Lit}, P}, ...
  ...
  A = bif:<boolean-op>(..., X, ...)
  B = succeeded,guard A
  br B, Success, Fail

If Lit is not 'true' nor 'false' we know that the succeed check will
fail, and thus change the branch from P to N to branch directly to
Fail. As this is in a guard context, we can be sure that the code we
short circuit doesn't have side effects nor defines anything used past
Fail.