beam_ssa_opt: Convert bool bifs to flow-control
Rewrite boolean bif:'and', bif:'and' and bif:'not' to use flow
control. In effect, this optimization converts 'and' and 'or' to be
short circuiting when it can be done safely.

We look for basic blocks which end with a conditional branch branching
on the result of the boolean bif. We only consider bifs which are not
followed by a succeeded instruction and where the conditional branch
is the only user.

Sinking of side-effect-free operations to after the first conditional
branch is mainly left for the ssa_opt_inter_block_sink pass. The
exception to this is the sinking of the bool bif's RHS when the bool
op is the only user and the def is in the same block as the bool
op. As ssa_opt_inter_block_sink does not touch anything inside a
receive this is necessary in order to turn some bifs into tests.