beam_ssa_opt: Only sink instructions when not inflating stack use
Update the inter-block sinking pass to only sink when we can guarantee
that it won't increase the stack use.

With this patch we only sink a def, `Def`, using the vars `Uses` from
the basic block `From` to the basic block `To` if one of the following
conditions hold:

* All members of `Uses` are live-in at `To`.

* Moving `Def` will not increase the spill set size at any
  x-clobbering instructions past `Def` in `From`.

* The net live-in size at `To` won't be increased by moving `Def`.

* The blocks through which we will extend the liveness ranges for
  `Uses` by moving `Def` from `From` to `To` contains no x-clobbering
  instructions.

When `Def` is inserted into `To` it is always inserted before any
x-clobbering instruction.

Future improvements are to update the liveness information so it won't
have to be recalculated for each iteration through the
ssa_opt_inter_block_sink1/1 loop.