digraph: Avoid false edges for vertices with names containing a '_'
This patch fixes a bug in the digraph module where in_edges/2 and
out_edges/2 return false edges in the case they are called for a
vertex which has a '_' in its name term. The following example shows
the problem:

    G = new([acyclic]),
    A = add_vertex(G, 'A'),
    B = add_vertex(G, '_'),
    AB = add_edge(G, A, B),
    0 = out_degree(G, B), %% ok
    [] = out_edges(G, B), %% Fails as [AB] is returned

An edge, E, emanating from and incident on a vertex is stored in the
ntab-ets table as a record of the form {{out, V}, E} and {{in, V}, E}
respectively. As out_edges/2 and in_edges/2 use ets:select/2 to look
up the relevant edges, a vertex with a '_' somewhere a in the term
used as its name will trigger ets to do a wildcard match. In the
example above, this will return all out-edges in the whole graph. This
problem is fixed by switching from ets:select to ets:lookup.

This patch includes an extended test case which checks the correctness
of the fix described above as well as checking that the other
edge-related functions in digraph do not suffer from the same problem.