compiler: Elide calls where the result is statically known
This patch adds a pass which first analyzes the first few basic blocks
of a function to determine if its result can be expressed as a simple
condition on its input parameters. The result of the analysis is then
used to, together with the results from the ssa_opt_type_continue
pass, elide calls to local functions while preserving the behavior of
the original program.