compiler gchains: Don't follow unconditional branches
Stop following unconditional branches when detecting guard chains as
it can lead to the fallthrough bypassing its true location and end up
in a block starting with a phi-node. This is a problem if, for
example, we are inserting the gchain profiling instruction as it will
be executed on unintended execution paths.