Generate multi-byte NOPs in the way GAS does
BeamAssembler::aligned_call() generates a nop (0x90) instruction for
each byte of alignment required. Most modern CPU:s are able to execute
up to four NOP instructions in a cycle [1] during ideal conditions,
but there are no guarantees for the instruction stream generated by
the BEAM loader to be ideal.

Looking at the BEAM code loaded by a typical run of the
`scripts/diffable` benchmark-tool, we find that there are in total
275086 locations requiring alignment. Of those, the number of
alignment bytes needed have the following distribution:

  bytes   percentage
      0    7.7%
      1    7.5%
      2    8.4%
      3   16.7%
      4   19.6%
      5    8.3%
      6    7.4%
      7   24.4%

Interesting to note is that 40% of the alignments will require, even
in ideal circumstances, more than once cycle to execute. Therefore
this changes the BeamAssembler::aligned_call() implementation to
generate multi-byte NOPs. The multi-byte NOPs used are the same as the
NOPs produced by the GNU Assembler when doing alignment in a code
segment.

[1] https://www.agner.org/optimize/