compiler: destructive-update: Don't force updated tuples onto heap
This patch removes the compiler's support for forcing literal tuples
which are destructively updated onto the heap. Destructive updates are
only legal to do when the updated value is on the heap. The
destructive-update pass ensures this by tracking where values are
coming from and, if the source is a literal, construct the values on
the heap. This used to be done for binaries in order to enable the use
of private_append and for tuples to use in-place updates.

In contrast to binaries which are boxed objects where the destructive
update is done in memory that is off-heap and which therefore doesn't
need a write-barrier to not violate garbage collector invariants,
tuples do. The write barrier used by destructive tuple updates
triggers a copy of the tuple is not on the young heap. As literals are
not on the young heap, the compiler does not need to force literal
tuples onto the heap. As the write-barrier test is always needed
before choosing the fast in-place update code path, there is no
drawback in not using put_tuple to construct what used to be literal
tuples, it might even be faster and provide a small decrease in code
size.

In addition to changing the compiler, this patch updates the test
suite and also extends it to cover compiler execution paths which the
existing tests no longer cover.