compiler: Remove dead code from destructive update pass
With the commit XXXXXXXXXXXX ("compiler: Use types to improve
destructive update pass") the destructive update pass is limited in
its exploration depth to the depth to which type information is
available (set by MAX_TYPE_DEPTH in beam_types.hrl). That the
exploration depth is now bounded allows the protection against
infinite loops as implemented in commit 96df0f1d5f80 ("compiler: Avoid
infinite loops in destructive update pass") to be removed as it now is
dead code.