compiler: ss: Don't include embeddings among aliasing edges
Don't include outgoing embedding edges among the aliasing edges as
they do not contribute to the alias status of the aggregate in which
the node is embedded.

This change corrects errors where aggregates were mistakenly
classified as aliased, but fortunately with regards to compiler
correctness this only leads to missed optimization opportunities.