compiler: Remove dead code from alias analysis pass
Remove an explicit error case in the alias analysis pass which
currently is the only uncovered line in the coverage report for the
pass. During development it was helpful to have a clear-text exit
instead of a match failure when an unknown instruction was
encountered, but now as the code has matured, it is only noise in the
coverage report. Therefore remove it, as, if a new ssa instruction is
added, it will be obvious from the error message what the problem is.