compiler alias analysis: Improve handling of Phi-instructions
Stop treating Phi instructions as an extraction, instead make use of
the infrastructure used to derive the structure and alias status of
function arguments.

The sharing state database representation for the result of a Phi
instruction is the same as for a function argument of a function with
multiple call sites, where the concrete argument values at the call
sites are the value at the respective predecessor block.