compiler alias analysis: Change work limiting strategy
Instead of limiting the maximum number of allowed traversals of the
functions in the module, limit the number of times a function is
analyzed before aborting the alias analysis pass.

Compared to the old limit strategy, this allows all more modules
compiled by diffable to successfully complete, without a noticeable
change in runtime.