compiler alias analysis: Fix bug in status of map extracts
As the destructive update pass does not have any support for patching
literal maps, as it can do for tuples to ensure that the term that
will be destructively updated is on the heap, we must conservatively
consider anything extracted from a literal map as aliased.

This is not a significant limitation as tracking the alias status of
individual associations of non-literal maps is currently considered
unfeasible.